Restructured types folder
